+++ src/main/java/fr/insalyon/creatis/moteurlite/boutiques/model/ValueRequires.java
@@ -1,5 +1,5 @@
 
-package fr.insalyon.creatis.moteurlite.boutiques.scheme;
+package fr.insalyon.creatis.moteurlite.boutiques.model;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 
 /**
- * Ids of the inputs that are disabled when the corresponding value choice is selected.
+ * Ids of the inputs that are required when the corresponding value choice is selected.
  * 
  */
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -20,7 +20,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 })
 @Generated("jsonschema2pojo")
-public class ValueDisables {
+public class ValueRequires {
 
     @JsonIgnore
     private Map<String, Object> additionalProperties = new LinkedHashMap<String, Object>();
