+++ src/main/java/fr/insalyon/creatis/moteurlite/runner/MoteurLiteRunner.java
@@ -22,9 +22,9 @@ import fr.insalyon.creatis.moteurlite.MoteurLite;
 import fr.insalyon.creatis.moteurlite.MoteurLiteConstants;
 import fr.insalyon.creatis.moteurlite.MoteurLiteException;
 import fr.insalyon.creatis.moteurlite.boutiques.BoutiquesService;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.BoutiquesDescriptor;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.Input;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.OutputFile;
+import fr.insalyon.creatis.moteurlite.boutiques.model.BoutiquesDescriptor;
+import fr.insalyon.creatis.moteurlite.boutiques.model.Input;
+import fr.insalyon.creatis.moteurlite.boutiques.model.OutputFile;
 import fr.insalyon.creatis.moteurlite.gasw.GaswMonitor;
 import fr.insalyon.creatis.moteurlite.gasw.WorkflowsDBRepository;
 import fr.insalyon.creatis.moteurlite.iteration.IterationService;
